refactor: 优化代码格式、setTimeout sleep、修正v-for循环没有key、拆分utils函数至更细颗粒度、优化 return 结构
