+++ src/blocks/table/table.vue
@@ -183,10 +183,7 @@ export default {
 
                 return this.handleTableColumn(this.block.options['table-column']);
             } else {
-
-                return this.handleTableColumn(
-                    Object.keys(this.fields).map(val => this.fields[val])
-                );
+                return this.handleTableColumn(Object.values(this.fields));
             }
         },
         expandRow() {
